+++ Changelog.md
@@ -16,6 +16,12 @@ and this project adheres to [Semantic Versioning].
 
 - [MainFrame] There won't be any expansion separators between the different S1 heroic dungeons anymore.
 
+- [SearchPanel&ApplicationViewer] Swapped out the RaiderIO data fortified and tyrannical timers for the new single ones.
+
+- [Lockouts] Lockouts will now be deleted when they're expired.
+
+- [SearchPanel] The raid category will now display "0/0" when there is no last season data.
+
 
 
 ## [2.5.3] - 2024-09-10